figure-6/requirements.txt
numpy==1.23.5
Pillow==9.3.0
scipy==1.9.3
scikit-image==0.19.3
pandas==1.5.2
r-scales==1.3.0
r-tidyr==1.3.1
r-dplyr==1.1.4
r-readr==2.1.5
r-purrr==1.0.2
r-stringi==1.8.4
r-stringr==1.5.1
r-forcats==1.0.0
r-ggpmisc==0.6.1
r-ggpp==0.5.8-1
r-cowplot==1.1.3
r-ggplot2==3.5.1
r-gridExtra==2.3
r-png==0.1-8
r-ggbeeswarm==0.7.2
